97: split out reducers to separate folder and moved files to core
