+++ client/src/containers/TeamDetails.js
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import { withRouter } from 'react-router'
 
 //Actions
-import * as actionCreators from '../redux/actions/actionCreators';
+import * as actionCreators from '../core/actions/actionCreators';
 
 import {
   TeamSummaryBox,
