+++ client/src/core/epics/index.js
@@ -1,49 +1,43 @@
 import { combineEpics } from 'redux-observable';
-import { combineReducers } from 'redux';
-import { routerReducer } from 'react-router-redux';
 
-import { reducer as formReducer } from 'redux-form';
 /*
   import epics and reducers here,
   they should be kept in the same file
   i.e. => import blahReducer, { blahEpic } from './blah'
 */
-import authReducer, { authEpic } from './auth';
-import onBoardingReducer, { onBoardingEpic } from './onBoarding';
-import userReducer, { getUserEpic, addUserEpic, updateUserEpic } from './user';
-import skills, {
+import { getUserEpic, addUserEpic, updateUserEpic } from './user';
+import {
   getSkillsEpic,
   addSkillEpic,
   deleteSkillEpic,
 } from './skills';
 
-import certifications, {
+import {
   getCertificationsEpic,
   addCertificationEpic,
   deleteCertificationEpic,
   editCertificationEpic,
 } from './certifications';
 
-import experiences, {
+import {
   getExperiencesEpic,
   addExperienceEpic,
   deleteExperienceEpic,
 }  from './experiences';
 
-import teams, {
+import {
   addTeamEpic,
   getTeamEpic,
   getAllTeamsEpic,
 }  from './team';
 
-import projects, {
+import {
   getProjectEpic,
   getAllProjectsEpic,
 }  from './project';
 
 export const rootEpic = combineEpics(
   /* epics */
-  // authEpic,
   getUserEpic,
   getSkillsEpic,
   addSkillEpic,
@@ -63,17 +57,3 @@ export const rootEpic = combineEpics(
   getProjectEpic,
   getAllProjectsEpic,
 );
-
-export const rootReducer = combineReducers({
-  /* reducers */
-  auth: authReducer,
-  onBoardingReducer,
-  user: userReducer,
-  router: routerReducer,
-  form: formReducer,
-  skills,
-  certifications,
-  experiences,
-  projects,
-  teams,
-})
