+++ client/src/core/epics/project.js
@@ -1,5 +1,5 @@
-import { getProjectFulfilled } from './../actions/actionCreators';
-import * as api from '../../core/api';
+import { getProjectFulfilled } from '../actions/actionCreators';
+import * as api from '../api/index';
 import {
   GET_PROJECT,
   GET_ALL_PROJECTS,
@@ -34,20 +34,7 @@ const getAllProjectsEpic = (action$) => {
     .concat(action$.mapTo({ type: GET_PROJECT_FULFILLED }))
 };
 
-const projects = (state = {
-  allProjects: [],
-}, action) => {
-  const { type, payload } = action;
-  switch (type) {
-    case GET_PROJECT_FULFILLED:
-      return {...state, allProjects: payload || []};
-    default:
-      return state;
-  }
-}
-
 export {
-  projects as default,
   getProjectEpic,
   getAllProjectsEpic,
 }
