+++ client/src/core/epics/skills.js
@@ -1,9 +1,8 @@
-import { getSkillsFullfilled, getSkills } from './../actions/actionCreators';
-import * as api from '../../core/api';
+import { getSkillsFullfilled, getSkills } from '../actions/actionCreators';
+import * as api from '../api/index';
 import {
   GET_SKILLS,
   GET_SKILLS_FULLFILLED,
-  SKILLS_WERE_CHECKED,
   ADD_SKILL,
   DELETE_SKILL,
 } from '../../utils/types';
@@ -47,23 +46,7 @@ const deleteSkillEpic = (action$) => {
     )
 };
 
-const skills = (state = {
-  list: [],
-  checked: false
-}, action) => {
-  const { type, payload } = action;
-  switch (type) {
-    case GET_SKILLS_FULLFILLED:
-      return Object.assign({}, state, { list: payload || [] });
-    case SKILLS_WERE_CHECKED:
-      return Object.assign({}, state, { checked: true });
-    default:
-      return state;
-  }
-};
-
 export {
-  skills as default,
   getSkillsEpic,
   addSkillEpic,
   deleteSkillEpic,
