+++ client/src/core/reducers/certifications.js
@@ -1,0 +1,15 @@
+import {
+  GET_CERTIFICATIONS_FULFILLED,
+} from '../../utils/types';
+
+export function certificationsReducer(state = {
+  list: [],
+}, action) {
+  const { type, payload } = action;
+  switch (type) {
+    case GET_CERTIFICATIONS_FULFILLED:
+      return {...state, list: payload };
+    default:
+      return state;
+  }
+}
