+++ client/src/core/reducers/onBoarding.js
@@ -1,0 +1,26 @@
+import { UPDATE_USER } from '../../utils/types'
+
+export function onBoardingReducer(state = {
+  currentPage: 1,
+  firstName: '',
+  lastName: '',
+  email: '',
+  date: undefined,
+  location: {
+
+  },
+  skills: {
+
+  },
+}, { type, payload}) {
+  switch (type) {
+    case(UPDATE_USER):
+      const { name, value } = payload;
+      return {
+        ...state,
+        [name]: value,
+      };
+    default:
+      return state
+  }
+}
